+++ CHANGELOG.md
@@ -10,3 +10,11 @@ and this project adheres to
 ## [0.0.1] - Unreleased
 
 ### Added
+
+- The option `-c` to connect to a specific Wireguar configuration
+- The option `-d` to disconnect to a specific Wireguar configuration
+- The option `-l` to list all Wireguard configuration
+- The option `-s` to show the status of Wireguard
+- The autocomplation for bash
+- The installer script
+- The `.wgbconf.json` configuration file